update cmake structure
