+++ c/conanfile.txt
@@ -1,5 +1,6 @@
 [requires]
 jsmn/1.1.0
+libunistring/1.1
 
 [generators]
 CMakeDeps
